backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
APScheduler==3.11.1
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
dnspython==2.8.0
fastapi==0.122.0
flatbuffers==25.9.23
h11==0.16.0
h3==4.3.1
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
mongomock==4.3.0
numpy==2.3.5
packaging==25.0
pluggy==1.6.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pymongo==4.15.4
pytest==9.0.1
pytest-asyncio==1.3.0
python-dotenv==1.2.1
python-geohash==0.8.5
pytz==2025.2
requests==2.32.5
respx==0.22.0
sentinels==1.1.1
sniffio==1.3.1
starlette==0.50.0
timezonefinder==8.1.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.38.0
